Changed camera prefab and added namespace to script
