--- Assets/CameraHandler.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 01f65c4e5e0a6fa458ed320993374618
