+++ Assets/Prefabs/CameraSystem.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fd0283a78bcb8c9458f6c5b873ee32d3
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
